refactored chat page and invitations page
