+++ apps/sim/app/chat/[subdomain]/page.tsx
@@ -1,4 +1,4 @@
-import ChatClient from '@/app/chat/[subdomain]/chat-client'
+import ChatClient from '@/app/chat/[subdomain]/chat'
 
 export default async function ChatPage({ params }: { params: Promise<{ subdomain: string }> }) {
   const { subdomain } = await params
