+++ apps/sim/app/chat/components/auth/email/email-auth.tsx
