+++ apps/sim/app/chat/components/auth/password/password-auth.tsx
