+++ apps/sim/app/chat/components/header/header.tsx
