+++ apps/sim/app/chat/components/index.ts
