+++ apps/sim/app/chat/components/input/voice-input.tsx
