+++ apps/sim/app/chat/components/loading-state/loading-state.tsx
