+++ apps/sim/app/chat/components/message-container/message-container.tsx
@@ -3,10 +3,7 @@
 import { memo, type RefObject } from 'react'
 import { ArrowDown } from 'lucide-react'
 import { Button } from '@/components/ui/button'
-import {
-  type ChatMessage,
-  ClientChatMessage,
-} from '@/app/chat/[subdomain]/components/message/message'
+import { type ChatMessage, ClientChatMessage } from '@/app/chat/components/message/message'
 
 interface ChatMessageContainerProps {
   messages: ChatMessage[]
