+++ apps/sim/app/chat/components/voice-interface/components/particles.tsx
