+++ apps/sim/app/chat/components/voice-interface/voice-interface.tsx
@@ -5,7 +5,7 @@ import { Mic, MicOff, Phone } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { createLogger } from '@/lib/logs/console/logger'
 import { cn } from '@/lib/utils'
-import { ParticlesVisualization } from '@/app/chat/[subdomain]/components/voice-interface/components/particles'
+import { ParticlesVisualization } from '@/app/chat/components/voice-interface/components/particles'
 
 const logger = createLogger('VoiceInterface')
