+++ apps/sim/app/chat/hooks/index.ts
