+++ apps/sim/app/invite/components/index.ts
@@ -1,0 +1,2 @@
+export { InviteLayout } from './layout'
+export { InviteStatusCard } from './status-card'
